requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# HTTP client
httpx==0.28.1

# Scheduler
apscheduler==3.10.4

# Google Gemini (sentiment analysis)
google-genai==1.5.0

# Dashboard
streamlit==1.41.1

# Configuration
python-dotenv==1.0.1

# Utilities
pydantic==2.10.4
